backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
alembic==1.15.2
python-dotenv==1.1.0
celery==5.3.6
redis==5.0.1
requests==2.31.0
openai==1.12.0
pillow==10.2.0
httpx==0.27.0
beautifulsoup4==4.12.3
markdown==3.6
python-docx==1.1.0 